Day 18 part II
This commit adds two new functions to the CustomLib class: CalculatePolygonArea and CalculatePolygonPerimeter. These functions take a vector of points as input and calculate the area and perimeter of the corresponding polygon. The area is calculated using the shoelace formula, while the perimeter is calculated by summing up the absolute differences between consecutive points.